rts54hub: Allow updating the RTD21xx devices in background mode
Original code by Ricky Wu <ricky_wu@realtek.com>
