+++ plugins/rts54hub/fu-plugin-rts54hub.c
@@ -9,7 +9,8 @@
 #include "fu-plugin-vfuncs.h"
 
 #include "fu-rts54hub-device.h"
-#include "fu-rts54hub-rtd21xx-device.h"
+#include "fu-rts54hub-rtd21xx-background.h"
+#include "fu-rts54hub-rtd21xx-foreground.h"
 
 void
 fu_plugin_init (FuPlugin *plugin)
@@ -17,7 +18,8 @@ fu_plugin_init (FuPlugin *plugin)
 	FuContext *ctx = fu_plugin_get_context (plugin);
 	fu_plugin_set_build_hash (plugin, FU_BUILD_HASH);
 	fu_plugin_add_device_gtype (plugin, FU_TYPE_RTS54HUB_DEVICE);
-	fu_plugin_add_device_gtype (plugin, FU_TYPE_RTS54HUB_RTD21XX_DEVICE);
+	fu_plugin_add_device_gtype (plugin, FU_TYPE_RTS54HUB_RTD21XX_BACKGROUND);
+	fu_plugin_add_device_gtype (plugin, FU_TYPE_RTS54HUB_RTD21XX_FOREGROUND);
 	fu_context_add_quirk_key (ctx, "Rts54TargetAddr");
 	fu_context_add_quirk_key (ctx, "Rts54I2cSpeed");
 	fu_context_add_quirk_key (ctx, "Rts54RegisterAddrLen");
