+++ plugins/rts54hub/meson.build
@@ -12,6 +12,8 @@ shared_module('fu_plugin_rts54hub',
   sources : [
     'fu-rts54hub-device.c',
     'fu-rts54hub-rtd21xx-device.c',
+    'fu-rts54hub-rtd21xx-background.c',
+    'fu-rts54hub-rtd21xx-foreground.c',
     'fu-plugin-rts54hub.c',
   ],
   include_directories : [
